trigger: Use get_dwarf_{arg,ret}spec()
When -A option is given with no explicit argspec, consult DWARF debug
info to retrieve the argument spec.  This enables more easier way to
specify arguments.

  $ gcc -pg -g -o args args.c

  $ uftrace -A main args
  # DURATION    TID     FUNCTION
              [12836] | main(1, 0x7fff587e86b8) {
     0.300 us [12836] |   foo();
     0.093 us [12836] |   bar();
     0.076 us [12836] |   null();
     3.419 us [12836] | } /* main */

  $ uftrace -A . args
  # DURATION    TID     FUNCTION
              [12842] | main(1, 0x7fff57933758) {
     0.265 us [12842] |   foo(-1, 32768);
     0.533 us [12842] |   bar("string argument", 'c', 0.000010);
     0.094 us [12842] |   null(NULL);
     3.674 us [12842] | } /* main */

TODO: It needs to save the generated argspec somewhere so that it can
apply them when replay (it parses the DWARF again at replay for now).

Signed-off-by: Namhyung Kim <namhyung@gmail.com>