arch/x86: Support -fno-plt libcall tracing
GCC has an option to suppress PLT for security reason.  With this option
.plt section has no entry (or even the section does not exist) and
libcalls are converted to indirect jumps from the GOT.

As old code missed some condition to check it properly, it was reported
that there's a segfault due to invalid PLT handling.  This code
robustify it at least for x86_64.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>