script: Pass script info to uftrace_begin callback
In order to be a more effective script, it needs to know some
information regarding the execution context.  To do that, it passes an
argument to the uftrace_begin() callback.  The argument has two info -
whether it's called from record and uftrace version.  It'll be added
later if needed.

Please note that it'll make the script not backward-compatible anymore.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>