utils: Introduce strv utility functions
The struct strv keeps a vector of strings terminated by an additional
NULL pointer like argv.  The strv_split() is to build a strv from a
given string separated by the delim.  The strv_join() does the
opposite.  The strv_for_each() macro can be used to iterate the strings
in the strv.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>