tui: Add more graph navigation keys
The 'n' key is to move to next sibling (i.e. one has same parent) and
the 'p' key is to move to previous sibling.  And the 'u' key is to move
to the parent.

For report mode, the 'u' key is ignored and other keys act as arrow keys.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>